e2e: move httpexpect Client to its own util module
Extracting generic httpexpect functionality from forwarder tests makes it usable by other projects.
The Client will now take http.RoundTripper, which can be fully customized.
